layer_v3/pyproject.toml
[tool.poetry]
name = "aws-lambda-powertools-python-layer"
version = "3.0.0"
description = "Powertools for AWS Lambda (Python) Lambda Layers"
authors = ["Powertools for AWS Maintainers <aws-powertools-maintainers@amazon.com>"]

license = "MIT"
[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = "^2.161.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
boto3 = "^1.24.46"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
